award/rejection fixes

+++ app/models/rejection.js
@@ -1,8 +1,8 @@
 module.exports = function (sequelize, DataTypes) {
 
 	var Rejection = sequelize.define('Rejection', {
-		ProposalId: 'MEDIUMINT',
-		YearId: 'MEDIUMINT',
+		ProposalId: 'INT',
+		YearId: 'INT',
 		Notes: 'MEDIUMTEXT'
 	});
